fkn pyton2

+++ projects/amkid_spikedetect.py
@@ -159,7 +159,7 @@ def findspikes_IQBT(windowtime=10., sig=4.5, expspikefree=75., crosstones=20., i
     # spike (say, it could be the wirescanner!)
     for windowidx in range(len(windows_tstart)):
         flaggedtones_thiswindow = np.sum(windowflag[windowidx, :])
-        print("DEBUG: WIN=%i | NFLAGGED=%.3f | NTHRESH=%.3f"%(windowidx, flaggedtones_thiswindow, crosstones/100 * nused))
+        print("DEBUG: WIN=%i | NFLAGGED=%i | NTHRESH=%.3f"%(windowidx, flaggedtones_thiswindow, int(crosstones/100. * float(nused))))
         if flaggedtones_thiswindow <= int(crosstones/100. * float(nused)):
             # Not real spike
             windowflag[windowidx, :] = False
